I wish I had any idea of what I'm doing

+++ src/j_cast.c
@@ -6,7 +6,7 @@
 /*   By: jmeier <jmeier@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/28 10:59:06 by jmeier            #+#    #+#             */
-/*   Updated: 2017/12/10 20:30:29 by jmeier           ###   ########.fr       */
+/*   Updated: 2017/12/10 20:36:43 by jmeier           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ void	j_int_cast(va_list arg)
 	intmax_t	output;
 
 	output = (intmax_t)va_arg(arg, void*);
-	ft_putnbr_base_fd((int)output, 10, 1);
+	ft_putnbr_base_fd(output, 10, 1);
 }
 
 void	j_oct_cast(va_list arg)
@@ -25,7 +25,7 @@ void	j_oct_cast(va_list arg)
 	uintmax_t	output;
 
 	output = (uintmax_t)va_arg(arg, void*);
-	ft_putunbr_base((unsigned int)output, 8);
+	ft_putunbr_base(output, 8);
 }
 
 void	j_hex_cast(va_list arg)
@@ -33,7 +33,7 @@ void	j_hex_cast(va_list arg)
 	uintmax_t	output;
 
 	output = (uintmax_t)va_arg(arg, void*);
-	ft_putunbr_base((unsigned int)output, 16);
+	ft_putunbr_base(output, 16);
 }
 
 void	j_lhex_cast(va_list arg)
@@ -41,7 +41,7 @@ void	j_lhex_cast(va_list arg)
 	uintmax_t	output;
 
 	output = (uintmax_t)va_arg(arg, void*);
-	ft_putunbr_base_lc((unsigned int)output, 16, 1);
+	ft_putunbr_base_lc(output, 16, 1);
 }
 
 void	j_uint_cast(va_list arg)
@@ -49,5 +49,5 @@ void	j_uint_cast(va_list arg)
 	uintmax_t	output;
 
 	output = (uintmax_t)va_arg(arg, void*);
-	ft_putulonglong((unsigned long long)output, 10, 1);
+	ft_putulonglong(output, 10, 1);
 }
